Time: 1 ms (27.27%) | Memory: 41.5 MB (8.69%) - LeetSync

+++ 342-power-of-four/power-of-four.java
@@ -1,6 +1,6 @@
 class Solution {
     public boolean isPowerOfFour(int n) {
-           for (double i = 0; i < 31; i++) {
+           for (int i = 0; i < 31; i++) {
             double ans =  Math.pow(4, i);
             if (ans == n) {
                 return true;
